put CPMLDR3.COM, CPMX.SYS and TESTCPMX.COM also on GitHub
Use TESTCPMX.COM for 1st test.
If it works, rename CPMX.SYS to CPM3.SYS (on A:).
Or rename CPMLDR3.COM to CPMLDR.COM and put it on root of the SD card.
It checks the AUTOEXEC flag and loads CPM3.SYS (=0) or CPMX.SYS (=1).
This gives a fallback to the original BIOS.

Signed-off-by: Martin <Ho-Ro@users.noreply.github.com>